+++ Lecture_10_Threads_CountDownLatch_Semaphore/pom.xml
@@ -5,7 +5,7 @@
     <artifactId>my-app</artifactId>
     <packaging>jar</packaging>
     <version>1.0-SNAPSHOT</version>
-    <name>my-app-9</name>
+    <name>my-app-10</name>
     <url>http://maven.apache.org</url>
 
     <properties>
