+++ Lecture_11_Barrier_CAS/pom.xml
@@ -1,11 +1,11 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.mycompany.streams</groupId>
-    <artifactId>my-app-streams</artifactId>
+    <groupId>com.mycompany.app</groupId>
+    <artifactId>my-app</artifactId>
     <packaging>jar</packaging>
     <version>1.0-SNAPSHOT</version>
-    <name>my-app-19</name>
+    <name>my-app-11</name>
     <url>http://maven.apache.org</url>
 
     <properties>
@@ -46,7 +46,7 @@
                 <finalName>${artifactId}-${version}</finalName>
 
                 <!--<resources>-->
-                    <!--<computation>-->
+                    <!--<resource>-->
                         <!--<targetPath>META-INF/plexus</targetPath>-->
                         <!--<filtering>false</filtering>-->
                         <!--<directory>${basedir}/src/main/plexus</directory>-->
@@ -56,7 +56,7 @@
                         <!--<excludes>-->
                             <!--<exclude>**/*.properties</exclude>-->
                         <!--</excludes>-->
-                    <!--</computation>-->
+                    <!--</resource>-->
                 <!--</resources>-->
 
             </build>
@@ -115,25 +115,19 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
+            <version>3.8.1</version>
+            <scope>test</scope>
+            <exclusions>
+                <exclusion>
+                    <groupId>*</groupId>
+                    <artifactId>*</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
             <version>4.12</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.glassfish</groupId>
-            <artifactId>javax.json</artifactId>
-            <version>1.0.4</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-lang3</artifactId>
-            <version>3.5</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-            <version>23.0</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
