+++ Lecture_11_Barrier_CAS/src/main/java/com/epam/executors/addition/prepare/util/Util.java
