--- Lecture_11_Barrier_ScheduledThreadPool/pom.xml
@@ -1,134 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <groupId>com.mycompany.app</groupId>
-    <artifactId>my-app</artifactId>
-    <packaging>jar</packaging>
-    <version>1.0-SNAPSHOT</version>
-    <name>my-app-10</name>
-    <url>http://maven.apache.org</url>
-
-    <properties>
-        <maven.compiler.source>1.7</maven.compiler.source>
-        <maven.compiler.target>1.7</maven.compiler.target>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.plugin.prop>2.2</maven.plugin.prop>
-        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    </properties>
-
-
-    <!-- BaseLine build-->
-    <build>
-        <defaultGoal>install</defaultGoal>
-        <directory>${basedir}/target</directory>
-        <finalName>${artifactId}-${version}</finalName>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <source>8</source>
-                    <target>8</target>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-
-
-    <profiles>
-        <profile>
-            <id>active</id>
-            <!-- Tomcat plugin -->
-            <!-- mvn org.apache.tomcat.maven:tomcat7-maven-plugin:2.2:run -->
-            <build>
-                <defaultGoal>package</defaultGoal>
-                <directory>${basedir}/build</directory>
-                <finalName>${artifactId}-${version}</finalName>
-
-                <!--<resources>-->
-                    <!--<resource>-->
-                        <!--<targetPath>META-INF/plexus</targetPath>-->
-                        <!--<filtering>false</filtering>-->
-                        <!--<directory>${basedir}/src/main/plexus</directory>-->
-                        <!--<includes>-->
-                            <!--<include>configuration.xml</include>-->
-                        <!--</includes>-->
-                        <!--<excludes>-->
-                            <!--<exclude>**/*.properties</exclude>-->
-                        <!--</excludes>-->
-                    <!--</resource>-->
-                <!--</resources>-->
-
-            </build>
-
-
-            <!-- NOTE: The release profile will be removed from future versions of the super POM -->
-        </profile>
-
-        <!-- NOTE: The release profile will be removed from future versions of the super POM -->
-        <profile>
-            <id>release-profile</id>
-            <!-- Tomcat plugin -->
-            <!-- mvn org.apache.tomcat.maven:tomcat7-maven-plugin:2.2:run -->
-            <build>
-                <pluginManagement>
-                    <plugins>
-                        <plugin>
-                            <groupId>org.apache.tomcat.maven</groupId>
-                            <artifactId>tomcat7-maven-plugin</artifactId>
-                            <version>${maven.plugin.prop}</version>
-                            <configuration>
-                                <url>http://localhost:8080/manager/text</url>
-                            </configuration>
-                        </plugin>
-
-                        <plugin>
-                            <artifactId>maven-antrun-plugin</artifactId>
-                            <version>1.1</version>
-                            <executions>
-                                <execution>
-                                    <id>echodir</id>
-                                    <goals>
-                                        <goal>run</goal>
-                                    </goals>
-                                    <phase>verify</phase>
-                                    <inherited>false</inherited>
-                                    <configuration>
-                                        <tasks>
-                                            <echo>Build Dir: ${project.build.directory}</echo>
-                                        </tasks>
-                                    </configuration>
-                                </execution>
-                            </executions>
-
-                        </plugin>
-                    </plugins>
-                </pluginManagement>
-            </build>
-
-
-            <!-- NOTE: The release profile will be removed from future versions of the super POM -->
-        </profile>
-    </profiles>
-
-    <dependencies>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>3.8.1</version>
-            <scope>test</scope>
-            <exclusions>
-                <exclusion>
-                    <groupId>*</groupId>
-                    <artifactId>*</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>4.12</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-</project>
