+++ Lecture_12_Benchmark_ForkJoinPool_Lambda_Java8/pom.xml
@@ -5,7 +5,7 @@
     <artifactId>my-app-threadpools</artifactId>
     <packaging>jar</packaging>
     <version>1.0-SNAPSHOT</version>
-    <name>my-app-11</name>
+    <name>my-app-12</name>
     <url>http://maven.apache.org</url>
 
     <properties>
