--- Lecture_13_Atomics/src/main/java/prepare/util/Utils.java
@@ -1,11 +1,0 @@
-package prepare.util;
-
-public class Utils {
-    public static void sleep(final int time) {
-        try {
-            Thread.sleep(time);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-    }
-}
