--- Lecture_13_Atomics/test/file.txt
@@ -1,5 +1,0 @@
-New file.txt
-Feature: Add Line 2
-Feature: Add Line 3
-Master: Add new Line 1Measter: add some info
-Master: Add new Line 1Add cherry-pick changes
