+++ Lecture_17_Streams_2_Optional_Builder_Creational/pom.xml
@@ -5,7 +5,7 @@
     <artifactId>my-app-streams</artifactId>
     <packaging>jar</packaging>
     <version>1.0-SNAPSHOT</version>
-    <name>my-app-19</name>
+    <name>my-app-17</name>
     <url>http://maven.apache.org</url>
 
     <properties>
