+++ Lecture_18_Streams_1_FunctionalInterface_Functions/pom.xml
@@ -5,7 +5,7 @@
     <artifactId>my-app-streams</artifactId>
     <packaging>jar</packaging>
     <version>1.0-SNAPSHOT</version>
-    <name>my-app-17</name>
+    <name>my-app-18</name>
     <url>http://maven.apache.org</url>
 
     <properties>
