+++ Lecture_9_Lambda_Java8_Addition/pom.xml
@@ -5,7 +5,7 @@
     <artifactId>my-app</artifactId>
     <packaging>jar</packaging>
     <version>1.0-SNAPSHOT</version>
-    <name>my-app-6</name>
+    <name>my-app-9</name>
     <url>http://maven.apache.org</url>
 
     <properties>
