+++ scripts/run-tests.sh
@@ -13,23 +13,7 @@ cd $TRAVIS_BUILD_DIR/Lecture_8_FixedSizeThreadPool_CachedThreadPool_Scheaduled_F
 mvn test
 cd $TRAVIS_BUILD_DIR/Lecture_9_Lambda_Java8_Addition
 mvn test
-cd $TRAVIS_BUILD_DIR/Lecture_10_Threads_CountDownLatch_ReentrantLock_Condition
+cd $TRAVIS_BUILD_DIR/Lecture_10_Threads_CountDownLatch_Semaphore
 mvn test
-cd $TRAVIS_BUILD_DIR/Lecture_11_Barrier_ScheduledThreadPool
-mvn test
-cd $TRAVIS_BUILD_DIR/Lecture_12_Benchmark_ForkJoinPool_Lambda_Java8
-mvn test
-cd $TRAVIS_BUILD_DIR/Lecture_13_Atomics_Locks_Synhronizers_ReadWriteLock_Semaphore_PartII_CAS_CompletableFuture
-mvn test
-cd $TRAVIS_BUILD_DIR/Lecture_14_Completable_Future_Concurrency_practice
-mvn test
-cd $TRAVIS_BUILD_DIR/Lecture_15_Concurrent_Collection
-mvn test
-cd $TRAVIS_BUILD_DIR/Lecture_16_Concurrent_Collection_Sync
-mvn test
-cd $TRAVIS_BUILD_DIR/Lecture_17_Streams_2_Optional_Builder_Creational
-mvn test
-cd $TRAVIS_BUILD_DIR/Lecture_18_Streams_1_FunctionalInterface_Functions
-mvn test
-cd $TRAVIS_BUILD_DIR/Lecture_19_Streams_BasicOperations
+cd $TRAVIS_BUILD_DIR/Lecture_11_Barrier_CAS
 mvn test
